model booking altered and appointment  history displayed as myappointments

+++ home/views.py
@@ -16,6 +16,8 @@ from .models import *
 
 from django.core.mail import send_mail
 from django.conf import settings
+from django.template.loader import render_to_string
+from django.core.mail import EmailMessage,EmailMultiAlternatives
 
   
 def register(request):
@@ -30,9 +32,19 @@ def register(request):
         user = form.cleaned_data.get('username')
         
         subject = f'MEDICAL APPOINTMENT ACCOUNT REGISTRATION'
-        message = 'Thank you {user} for registering! Your account has been successfully created. You can now log in and start exploring our services. If you have any questions or need assistance, feel free to reach out to our support team'
+        message = 'Account Register Confirmation '
         recepiant = form.cleaned_data.get('email')
-        send_mail(subject, message,settings.EMAIL_HOST_USER,[recepiant],fail_silently=False)
+        user = form.cleaned_data.get  ('username')
+        context ={
+          'user' : user
+        }
+        card_html = render_to_string('registeremail.html', context)
+        
+        email = EmailMultiAlternatives(subject, message,settings.EMAIL_HOST_USER, [recepiant])
+     
+        email.attach_alternative(card_html, 'text/html')
+        email.send(fail_silently=False)
+        
         messages.success(request,'success')
       
         messages.success(request,'Account created for '+ user)
@@ -40,7 +52,9 @@ def register(request):
   
   context ={ 'form' :form}
   return render(request,'register.html', context)
-  
+
+
+ 
   
   
 def user_login(request):
@@ -62,22 +76,31 @@ def user_login(request):
   context = {}
   return render(request,'login.html', context)
   
+
+
   
   
 def logout_user(request):
   logout(request)
   return redirect('login')
- 
+
+
+
 @login_required(login_url ='login') 
 def index(request):
   dict_dept = {
     'dept' : Department.objects.all()
   }
   return render(request, 'home.html',dict_dept)
-  
+
+
+ 
 @login_required(login_url ='login')   
 def about(request):
   return render(request, 'about.html')
+
+
+
 
 @login_required(login_url ='login') 
 def booking(request):
@@ -88,16 +111,14 @@ def booking(request):
   
   if request.method == 'POST':
     form = Bookingform(request.POST)
+    
     if form.is_valid():
+      form.instance.user = request.user
       form.save()
       patient = form.cleaned_data.get('p_name')
       booking_date = form.cleaned_data.get('booking_date')
       booking_time = form.cleaned_data.get('booking_time')
       doctor = form.cleaned_data.get('doc_name')
-      subject = 'MEDICAL APPOINTMENT'
-      message = f'Thank you { patient } for booking your medical appointment on { booking_date } at { booking_time } for  { doctor }. Please arrive 15 minutes early and remember to bring any necessary documentation or insurance cards. If you need to reschedule or have any questions, please contact us at medicalAppointment.com. We look forward to assisting you with your healthcare needs.'
-      recepiant = form.cleaned_data.get('p_email')
-      send_mail(subject, message,settings.EMAIL_HOST_USER,[recepiant],fail_silently=False)
       context = {
        'patient' : patient,
        'booking_time': booking_time,
@@ -105,25 +126,44 @@ def booking(request):
        'doctor': doctor
        
       }
+      card_html = render_to_string('emailsend.html', context)
+
+      subject = 'MEDICAL APPOINTMENT'
+      message = 'Medical Appointment confirmation'
+      recepiant = form.cleaned_data.get('p_email')
+      #send_mail(subject, message,settings.EMAIL_HOST_USER,[recepiant],html_message=card_html, content_type='text/html',fail_silently=False)
+      email = EmailMultiAlternatives(subject, message,settings.EMAIL_HOST_USER, [recepiant])
      
-      return render(request,'confirmation.html', context)
+      email.attach_alternative(card_html, 'text/html')
+      email.send(fail_silently=False)
+        
+      return render(request,'confirmation.html',context)
   return render(request, 'booking.html',dict_form)
+ 
+
 
 @login_required(login_url ='login') 
 def contact(request):
   return render(request, 'contact.html')
  
+
+
  
 
 @login_required(login_url ='login') 
 def delete(request,pk):
   instance = Booking.objects.get(pk=pk)
-  instance.delete()
+  if request.method == 'POST':
+    instance.delete()
+  
   dict_book = {
     'booking' : Booking.objects.all()
   }
   return render(request,'appointments.html',dict_book)
- 
+
+
+
+
 @login_required(login_url ='login') 
 def appointments(request):
   dict_book = {
@@ -131,10 +171,21 @@ def appointments(request):
   }
   return render(request,'appointments.html',dict_book)
  
-  
+ 
+
+
 @login_required(login_url ='login') 
 def doctors(request):
   dict_dept = {
     'dept' : Doctor.objects.all()
   }
   return render(request, 'doctors.html',dict_dept)
+  
+  
+@login_required(login_url='login')
+
+def myappointments(request):
+    user = request.user  # Assuming you are using Django's built-in User model
+    user_bookings = Booking.objects.filter(user=user)
+
+    return render(request, 'myappointments.html', {'bookings': user_bookings})
